Create get_variable_protein_site_sub.py
This script generates a csv file. Each column is a non-synonymous protein substitution and its respective sequence position on a gene protein sequence, that is different from the other proteins in a gene from different individuals. Can also be described as protein variants. The input file for this python script is a MUSCLE (Edgar RC ) generated multiple sequence alignment file. This Python scripts made use of several libraries that I learned in class like Pandas, and sys and incorporates my prior experience with Perl bioseq functions part of the bioperl/biopython modules, getopts for positional parameters,and glob. It is the subsequence analysis of part of a workflow I am establishing that is comprised of Perl, Shell and now Python scripts, which generate these MUSCLE alignment files. 

Edgar RC (2004). "MUSCLE: a multiple sequence alignment method with reduced time and space complexity". BMC Bioinformatics